Safari-609.1.11.1/LayoutTests/fast/forms/file/entries-api/webkitdirectory-drag-folder-webkitEntries-expected.txt

Tests dragging a folder onto a file input with webkitdirectory set and querying input.webkitEntries.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS inputElement.webkitdirectory is true
PASS entries.length is 0
PASS files.length is 6
PASS successfullyParsed is true

TEST COMPLETE
